+++ libs/data-mapper-v2/src/ui/ReactFlow.tsx
@@ -1,14 +1,24 @@
 import type { AppDispatch, RootState } from '../core/state/Store';
-import { useEffect, useMemo, useRef, useCallback, useState, useLayoutEffect } from 'react';
+import { useEffect, useMemo, useRef, useCallback, useState, useLayoutEffect, type MouseEvent } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import type { Connection, Node, Edge, ConnectionLineComponent, NodeProps, NodeTypes, OnNodeDrag, IsValidConnection } from '@xyflow/react';
-import { ReactFlow, addEdge, useReactFlow } from '@xyflow/react';
+import type {
+  Connection,
+  Node,
+  Edge,
+  ConnectionLineComponent,
+  NodeProps,
+  NodeTypes,
+  OnNodeDrag,
+  IsValidConnection,
+  EdgeChange,
+} from '@xyflow/react';
+import { ReactFlow, addEdge, applyEdgeChanges, useEdges, useReactFlow } from '@xyflow/react';
 import { reactFlowStyle, useStaticStyles, useStyles } from './styles';
 import SchemaNode from '../components/common/reactflow/SchemaNode';
 import ConnectionLine from '../components/common/reactflow/ConnectionLine';
 import ConnectedEdge from '../components/common/reactflow/ConnectedEdge';
 import type { ConnectionAction } from '../core/state/DataMapSlice';
-import { updateFunctionPosition, makeConnectionFromMap, setSelectedItem } from '../core/state/DataMapSlice';
+import { updateFunctionPosition, makeConnectionFromMap, setSelectedItem, updateEdgePopOverId } from '../core/state/DataMapSlice';
 import { FunctionNode } from '../components/common/reactflow/FunctionNode';
 import { useDrop } from 'react-dnd';
 import useResizeObserver from 'use-resize-observer';
@@ -17,7 +27,8 @@ import { convertWholeDataMapToLayoutTree } from '../utils/ReactFlow.Util';
 import { createEdgeId } from '../utils/Edge.Utils';
 import useAutoLayout from './hooks/useAutoLayout';
 import cloneDeep from 'lodash/cloneDeep';
-
+import EdgePopOver from '../components/canvas/EdgePopOver';
+import { getReactFlowNodeId } from '../utils/Schema.Utils';
 interface DMReactFlowProps {
   setIsMapStateDirty?: (isMapStateDirty: boolean) => void;
   updateCanvasBoundsParent: (bounds: Bounds | undefined) => void;
@@ -29,18 +40,27 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
   const reactFlowInstance = useReactFlow();
   const ref = useRef<HTMLDivElement>(null);
   const dispatch = useDispatch<AppDispatch>();
-  const { sourceNodesMap, targetNodesMap, functionNodes, flattenedSourceSchema, flattenedTargetSchema, dataMapConnections } = useSelector(
-    (state: RootState) => state.dataMap.present.curDataMapOperation
-  );
+  const edges = useEdges();
+  const {
+    sourceNodesMap,
+    targetNodesMap,
+    functionNodes,
+    flattenedSourceSchema,
+    flattenedTargetSchema,
+    dataMapConnections,
+    sourceStateConnections,
+  } = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation);
   const [functionNodesForDragDrop, setFunctionNodesForDragDrop] = useState<Node[]>([]);
   const { width = undefined, height = undefined } = useResizeObserver<HTMLDivElement>({
     ref,
   });
+  const [edgePopoverBounds, setEdgePopoverBounds] = useState<Bounds>();
 
-  const edges: Edge[] = useMemo(() => {
+  const realEdges: Edge[] = useMemo(() => {
+    let edges: Edge[] = [];
     if (Object.entries(dataMapConnections).length > 0) {
       const layout = convertWholeDataMapToLayoutTree(flattenedSourceSchema, flattenedTargetSchema, functionNodes, dataMapConnections);
-      return layout.edges.map((edge) => {
+      edges = layout.edges.map((edge) => {
         const newEdge: Edge = {
           id: createEdgeId(edge.sourceId, edge.targetId),
           source: edge.sourceId,
@@ -54,8 +74,41 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
       });
     }
 
-    return [];
+    return edges;
   }, [dataMapConnections, flattenedSourceSchema, flattenedTargetSchema, functionNodes]);
+
+  // Edges created when node is expanded/Collapsed
+  const temporaryEdgesMap: Record<string, Edge> = useMemo(() => {
+    const newEdgesMap: Record<string, Edge> = {};
+    const sourceStateConnectionsEntries = Object.entries(sourceStateConnections);
+
+    if (sourceStateConnectionsEntries.length > 0) {
+      for (const entry of sourceStateConnectionsEntries) {
+        const sourceNodeId = getReactFlowNodeId(entry[0], true);
+        const targetIds = Object.keys(entry[1]);
+        for (const targetId of targetIds) {
+          const targetNodeId = getReactFlowNodeId(targetId, false);
+          const id = createEdgeId(sourceNodeId, targetNodeId);
+          const edge: Edge = {
+            id: id,
+            source: sourceNodeId,
+            target: targetNodeId,
+            type: 'connectedEdge',
+            reconnectable: 'target',
+            focusable: true,
+            deletable: true,
+            animated: true,
+            data: {
+              isTemporary: true,
+            },
+          };
+          newEdgesMap[id] = edge;
+        }
+      }
+    }
+
+    return newEdgesMap;
+  }, [sourceStateConnections]);
 
   useAutoLayout();
 
@@ -80,9 +133,39 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
         position: node[1].position || { x: 10, y: 200 },
         draggable: true,
         selectable: false,
+        measured: { width: 1, height: 1 },
       }))
     );
   }, [functionNodes]);
+
+  useLayoutEffect(() => {
+    const edgeChanges: Record<string, EdgeChange> = {};
+
+    for (const [id, edge] of Object.entries(temporaryEdgesMap)) {
+      edgeChanges[id] = {
+        type: 'add',
+        item: edge,
+      };
+    }
+
+    for (const edge of edges) {
+      if (edge.data?.isTemporary) {
+        const id = edge.id;
+        if (temporaryEdgesMap[id]) {
+          delete edgeChanges[id];
+        } else {
+          edgeChanges[id] = {
+            type: 'remove',
+            id: id,
+          };
+        }
+      }
+    }
+
+    if (Object.entries(edgeChanges).length > 0) {
+      applyEdgeChanges(Object.values(edgeChanges), edges);
+    }
+  }, [edges, temporaryEdgesMap]);
 
   const isMapStateDirty = useSelector((state: RootState) => state.dataMap.present.isDirty);
 
@@ -180,6 +263,22 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
     [dispatch]
   );
 
+  const onEdgeContextMenu = useCallback(
+    (e: MouseEvent, edge: Edge) => {
+      e.preventDefault();
+      e.stopPropagation();
+
+      dispatch(updateEdgePopOverId(edge.id));
+      setEdgePopoverBounds({
+        x: e.clientX,
+        y: e.clientY,
+        width: 5,
+        height: 5,
+      });
+    },
+    [dispatch, setEdgePopoverBounds]
+  );
+
   const nodes = useMemo(
     () => [...Object.values(sourceNodesMap), ...Object.values(targetNodesMap), ...functionNodesForDragDrop],
     [sourceNodesMap, targetNodesMap, functionNodesForDragDrop]
@@ -191,8 +290,8 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
         id="dm-react-flow"
         ref={drop}
         nodes={cloneDeep(nodes)}
-        edges={edges}
-        nodeDragThreshold={0}
+        edges={[...realEdges, ...Object.values(temporaryEdgesMap)]}
+        nodeDragThreshold={1}
         onlyRenderVisibleElements={false}
         zoomOnScroll={false}
         zoomOnPinch={false}
@@ -201,7 +300,7 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
         edgeTypes={edgeTypes}
         preventScrolling={false}
         minZoom={1}
-        elementsSelectable={true}
+        elementsSelectable={false}
         maxZoom={1}
         autoPanOnConnect={false}
         snapToGrid={true}
@@ -215,6 +314,7 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
         onNodeDrag={onFunctionNodeDrag}
         onNodeDragStop={onFunctionNodeDragStop}
         isValidConnection={isValidConnection}
+        onEdgeContextMenu={onEdgeContextMenu}
         onConnect={onEdgeConnect}
         connectionLineComponent={ConnectionLine as ConnectionLineComponent | undefined}
         elevateEdgesOnSelect={true}
@@ -235,6 +335,7 @@ export const DMReactFlow = ({ setIsMapStateDirty, updateCanvasBoundsParent }: DM
             : undefined
         }
       />
+      {edgePopoverBounds && <EdgePopOver {...edgePopoverBounds} />}
     </div>
   );
 };