+++ libs/designer-ui/src/lib/panel/panelcontainer.tsx
@@ -59,6 +59,7 @@ export const PanelContainer = ({
   setOverrideWidth,
   overrideWidth,
   isResizeable,
+  mountNode,
 }: PanelContainerProps) => {
   const intl = useIntl();
 
@@ -103,7 +104,7 @@ export const PanelContainer = ({
           canResubmit={canResubmit}
           onUnpinAction={canUnpin ? onUnpinAction : undefined}
           resubmitOperation={() => resubmitOperation?.(nodeId)}
-          commentChange={() => onCommentChange(nodeId)}
+          commentChange={(newValue) => onCommentChange(nodeId, newValue)}
           toggleCollapse={toggleCollapse}
           onTitleChange={onTitleChange}
           onTitleBlur={onTitleBlur}
@@ -181,10 +182,11 @@ export const PanelContainer = ({
       modalType="non-modal"
       mountNode={{
         className: 'msla-panel-host-container',
+        element: mountNode,
       }}
       open={true}
       position={isRight ? 'end' : 'start'}
-      style={{ width: drawerWidth }}
+      style={{ position: 'absolute', width: drawerWidth }}
     >
       {isEmptyPane || isCollapsed ? (
         <Button