+++ libs/designer/src/lib/ui/panel/connectionsPanel/createConnection/createConnection.tsx
@@ -56,6 +56,7 @@ export interface CreateConnectionProps {
   description?: string;
   identity?: ManagedIdentity;
   isLoading?: boolean;
+  createText?: string;
   createConnectionCallback?: (
     newName?: string,
     selectedParameterSet?: ConnectionParameterSet,
@@ -82,6 +83,7 @@ export interface CreateConnectionProps {
 export const CreateConnection = (props: CreateConnectionProps) => {
   const {
     classes,
+    createText,
     nodeIds = [],
     showActionBar = true,
     iconUri = '',
@@ -236,7 +238,7 @@ export const CreateConnection = (props: CreateConnectionProps) => {
   );
 
   const unfilteredParameters: Record<string, ConnectionParameterSetParameter | ConnectionParameter> = useMemo(
-    () => (isMultiAuth ? { ...multiAuthParams } : { ...singleAuthParams }) ?? {},
+    () => (isMultiAuth ? { ...multiAuthParams } : { ...singleAuthParams }),
     [isMultiAuth, multiAuthParams, singleAuthParams]
   );
 
@@ -300,7 +302,10 @@ export const CreateConnection = (props: CreateConnectionProps) => {
       isTenantServiceEnabled() &&
       usingAadConnection &&
       isUsingOAuth &&
-      Object.keys(connectionParameters ?? {}).some((key) => equals(key, SERVICE_PRINCIPLE_CONSTANTS.CONFIG_ITEM_KEYS.TOKEN_TENANT_ID)),
+      Object.keys(connectionParameters?.['token']?.oAuthSettings?.customParameters ?? {}).some((key: string) => equals(key, 'tenantId')) &&
+      Object.keys(connectionParameters ?? {}).some((key: string) =>
+        equals(key, SERVICE_PRINCIPLE_CONSTANTS.CONFIG_ITEM_KEYS.TOKEN_TENANT_ID)
+      ),
     [connectionParameters, isUsingOAuth, usingAadConnection]
   );
 
@@ -482,12 +487,13 @@ export const CreateConnection = (props: CreateConnectionProps) => {
     return '';
   }, [authDescriptionText, isUsingOAuth, parameters, simpleDescriptionText]);
 
+  const createConnectionText = createText ?? createButtonText;
   const submitButtonText = useMemo(() => {
     if (isLoading) {
       return isUsingOAuth ? signInButtonLoadingText : createButtonLoadingText;
     }
-    return isUsingOAuth ? signInButtonText : createButtonText;
-  }, [createButtonLoadingText, createButtonText, isLoading, isUsingOAuth, signInButtonLoadingText, signInButtonText]);
+    return isUsingOAuth ? signInButtonText : createConnectionText;
+  }, [createButtonLoadingText, createConnectionText, isLoading, isUsingOAuth, signInButtonLoadingText, signInButtonText]);
 
   const submitButtonAriaLabel = useMemo(() => {
     return isUsingOAuth ? signInButtonAria : createButtonAria;