+++ libs/logic-apps-shared/src/utils/src/lib/models/dropdownMenuCustomNode.ts
@@ -1,0 +1,4 @@
+export interface DropdownMenuCustomNode {
+  priority?: number;
+  renderCustomComponent: () => React.ReactNode;
+}